refactor(services): migrate swagger schemas to service domain directories
