+++ packages/services/src/did-manager/common/utils.test.ts
