+++ packages/services/src/did-manager/common/utils.ts
