+++ packages/services/src/did-manager/common/verify-did-web.ts
