+++ packages/services/src/did-manager/common/verify-did-webvh.ts
