+++ packages/services/src/did-manager/common/verify.test.ts
