+++ packages/services/src/did-manager/common/verify.ts
